Prototype release of InVehicleAgentDisplay
Added a page for the InVehicleAgentDisplay, the screen users will interact with in trials.

Todo:
-Implement jsx logic to call screens to inVechicleAgentDisplay throughout scenario
-Implement behind interactables on this screen to call video changing functions

limitations: The lab is not finished completing the scenarios for the pilot study yet so we are using place holders. This means the test logic implemented will only be to show off the features we developed to enable the test scenarios that will be given.