TrialUI under development
No more bug scenario on bootup. Starting trial is smooth and simple.
-Created generalized user interface with new visual design. Screens will be called into a primary UI model. Progress loading bar added to show time left in trip. Timer added to display time left in the trial.
-Date time added to top left.
-Made button row.
-Made Help button
What is left is to implement the primary box interface which will receive screen components. Screen components will contain the elements to display inside the box in grid format. Screen functions will be responsible for setting their name in the database. Setting the name on change triggers a push back on to an array for keeping track of screen sequences.