Google maps and Trial Screens
Created TrialScreenInformation.js which creates an easily callable jsx component which when given a string and an onClose function will provide the user with a basic informational screen popping up over the vehicle UI. This should make it easier for the lab to create new screens in the future.

Further building into the google maps api component.

Working on
-Calling an array of jsx components to create a generalized trial format
-Tracking screen information on open and on close which will be used to reboot trial from last saved point.
-Two more components for different trial screens
-Ending trial process.