Trial Pause functionality
Creates useTimeout hook which creates the desired timeout effect on each screen. This timeout will trigger the onClose function after a certain number of seconds has passed. It will pause when the rest of the system is paused. Pause functionality completed for scheduled and non-scheduled pauses. Next working on question logic for impact which screens show during trial